requirements.txt
langgraph
langchain-openai
python-dotenv
mcp
httpx
langchain-core
rich
fastapi
uvicorn
redis
pydantic
requests
pyyaml
supabase
markdown
